Moved the loading dialogue into the dialogues folder and adjusted it's namespace appropriatley.
